requirements.txt
crewai
langchain-google-genai>=1.0.3
fastapi
uvicorn
python-dotenv
google-generativeai>=0.8.3
langgraph
langchain-core
arxiv
